@@ -4,17 +4,16 @@
   "private": true,
   "dependencies": {
     "@types/node": "^15.14.9",
-    "@types/react": "^17.0.68",
-    "@types/react-dom": "^17.0.21",
     "@types/react-router-dom": "^5.3.3",
     "@types/styled-components": "^5.1.28",
-    "antd": "^4.24.14",
+    "antd": "^5.26.6",
     "i18next": "^19.9.2",
     "i18next-browser-languagedetector": "^6.1.8",
     "i18next-xhr-backend": "^3.2.2",
-    "react": "^18.2.0",
-    "react-awesome-reveal": "^3.8.1",
-    "react-dom": "^18.2.0",
+    "nodemailer": "^7.0.5",
+    "react": "^18.3.1",
+    "react-awesome-reveal": "^4.3.1",
+    "react-dom": "^18.3.1",
     "react-i18next": "^11.18.6",
     "react-router-dom": "^5.3.4",
     "react-scripts": "^5.0.1",
@@ -41,5 +40,9 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@types/react": "^18.3.23",
+    "@types/react-dom": "^18.3.7"
   }
 }
