@@ -1,7 +1,7 @@
 import { BrowserRouter } from "react-router-dom";
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client"; // Import createRoot from 'react-dom/client'
 import { I18nextProvider } from "react-i18next";
-import 'antd/dist/antd.min.css';
+import 'antd/dist/reset.css'; // Changed from 'antd/dist/antd.css' to 'antd/dist/reset.css'
 
 import Router from "./router";
 import i18n from "./translation";
@@ -14,4 +14,14 @@ const App = () => (
   </BrowserRouter>
 );
 
-ReactDOM.render(<App />, document.getElementById("root"));
+// Get the root element from the DOM
+const rootElement = document.getElementById("root");
+
+// Check if the root element exists before creating the root
+if (rootElement) {
+  // Create a root and render your App component
+  const root = ReactDOM.createRoot(rootElement);
+  root.render(<App />);
+} else {
+  console.error("Root element with ID 'root' not found in the document.");
+}
